requirements.txt
Appium-Python-Client==0.26
autopep8==1.3.4
behave==1.2.5
enum34==1.1.6
parse==1.8.2
parse-type==0.4.2
pycodestyle==2.3.1
PyHamcrest==1.9.0
selenium==3.8.1
six==1.11.0
